+++ src/main/java/io/github/randomusert/mods/gregtech_pack_core/compat/Gregtech_pack_coreJEIPlugin.java
@@ -1,0 +1,31 @@
+package io.github.randomusert.mods.gregtech_pack_core.compat;
+
+
+import io.github.randomusert.mods.gregtech_pack_core.Gregtech_pack_core;
+import io.github.randomusert.mods.gregtech_pack_core.item.ModItems;
+import mezz.jei.api.IModPlugin;
+import mezz.jei.api.JeiPlugin;
+import mezz.jei.api.constants.VanillaTypes;
+import mezz.jei.api.registration.IRecipeRegistration;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ItemStack;
+
+@JeiPlugin
+public class Gregtech_pack_coreJEIPlugin implements IModPlugin {
+
+    private static final ResourceLocation UID = ResourceLocation.fromNamespaceAndPath(Gregtech_pack_core.MODID, "gtpc_jei_plugin");
+    @Override
+    public ResourceLocation getPluginUid() {
+        return UID;
+    }
+
+    @Override
+    public void registerRecipes(IRecipeRegistration registration) {
+        registration.addIngredientInfo(
+                new ItemStack(ModItems.SINGULARITY_ALLOY.get()),
+                VanillaTypes.ITEM_STACK,
+                Component.translatable("jei.gtpc.items.singularity_alloy.info")
+        );
+    }
+}