+++ src/main/java/io/github/randomusert/mods/gregtech_pack_core/datagen/ModDatapackProvider.java
@@ -1,0 +1,26 @@
+package io.github.randomusert.mods.gregtech_pack_core.datagen;
+
+import io.github.randomusert.mods.gregtech_pack_core.Gregtech_pack_core;
+import io.github.randomusert.mods.gregtech_pack_core.worldgen.ModBiomeModifiers;
+import io.github.randomusert.mods.gregtech_pack_core.worldgen.ModConfiguredFeatures;
+import io.github.randomusert.mods.gregtech_pack_core.worldgen.ModPlacedFeatures;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.PackOutput;
+import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+
+public class ModDatapackProvider extends DatapackBuiltinEntriesProvider {
+    public static RegistrySetBuilder BUILDER = new RegistrySetBuilder()
+            .add(Registries.CONFIGURED_FEATURE, ModConfiguredFeatures::bootstrap)
+            .add(Registries.PLACED_FEATURE, ModPlacedFeatures::bootstrap)
+            .add(NeoForgeRegistries.Keys.BIOME_MODIFIERS, ModBiomeModifiers::bootstrap);
+
+    public ModDatapackProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {
+        super(output, registries, BUILDER, Set.of(Gregtech_pack_core.MODID));
+    }
+}