+++ src/main/java/io/github/randomusert/mods/gregtech_pack_core/worldgen/ModConfiguredFeatures.java
@@ -1,0 +1,49 @@
+package io.github.randomusert.mods.gregtech_pack_core.worldgen;
+
+import io.github.randomusert.mods.gregtech_pack_core.Gregtech_pack_core;
+import io.github.randomusert.mods.gregtech_pack_core.block.ModBlocks;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.worldgen.BootstrapContext;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
+import net.minecraft.world.level.levelgen.structure.templatesystem.BlockMatchTest;
+import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
+import net.minecraft.world.level.levelgen.structure.templatesystem.TagMatchTest;
+
+import java.util.List;
+
+public class ModConfiguredFeatures {
+    public static final ResourceKey<ConfiguredFeature<?, ?>> OVERWORLD_XENOVERD_ORE_KEY = registerKey("xenoverd_ore");
+    public static final ResourceKey<ConfiguredFeature<?, ?>> END_XENOVERD_ORE_KEY = registerKey("end_xenoverd_ore");
+
+    public static void bootstrap(BootstrapContext<ConfiguredFeature<?, ?>> context) {
+
+
+        RuleTest stoneReplacables = new TagMatchTest(BlockTags.STONE_ORE_REPLACEABLES);
+        RuleTest deepslateReplacables = new TagMatchTest(BlockTags.DEEPSLATE_ORE_REPLACEABLES);
+
+        List<OreConfiguration.TargetBlockState> overworldXenoverdOres = List.of(
+                OreConfiguration.target(stoneReplacables, ModBlocks.XENOVERD_ORE.get().defaultBlockState()),
+                OreConfiguration.target(deepslateReplacables, ModBlocks.XENOVERD_ORE.get().defaultBlockState()));
+
+        RuleTest endReplecable = new BlockMatchTest(Blocks.END_STONE);
+        register(context, OVERWORLD_XENOVERD_ORE_KEY, Feature.ORE, new OreConfiguration(overworldXenoverdOres, 2));
+        register(context, END_XENOVERD_ORE_KEY, Feature.ORE, new OreConfiguration(endReplecable, ModBlocks.XENOVERD_ORE.get().defaultBlockState(), 9));
+
+    }
+
+    public static ResourceKey<ConfiguredFeature<?, ?>> registerKey(String name) {
+        return ResourceKey.create(Registries.CONFIGURED_FEATURE, ResourceLocation.fromNamespaceAndPath(Gregtech_pack_core.MODID, name));
+    }
+
+    private static <FC extends FeatureConfiguration, F extends Feature<FC>> void register(BootstrapContext<ConfiguredFeature<?, ?>> context,
+                                                                                          ResourceKey<ConfiguredFeature<?, ?>> key, F feature, FC configuration) {
+        context.register(key, new ConfiguredFeature<>(feature, configuration));
+    }
+}