+++ src/main/java/io/github/randomusert/mods/gregtech_pack_core/worldgen/ModPlacedFeatures.java
@@ -1,0 +1,42 @@
+package io.github.randomusert.mods.gregtech_pack_core.worldgen;
+
+import io.github.randomusert.mods.gregtech_pack_core.Gregtech_pack_core;
+import net.minecraft.commands.arguments.HeightmapTypeArgument;
+import net.minecraft.core.Holder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.worldgen.BootstrapContext;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.levelgen.VerticalAnchor;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.placement.HeightRangePlacement;
+import net.minecraft.world.level.levelgen.placement.PlacedFeature;
+import net.minecraft.world.level.levelgen.placement.PlacementModifier;
+
+import java.util.List;
+
+public class ModPlacedFeatures {
+
+    public static final  ResourceKey<PlacedFeature> XENOVERD_ORE_PLACED_KEY = registerKey("xenoverd_ore_placed");
+    public static final  ResourceKey<PlacedFeature> END_XENOVERD_ORE_PLACED_KEY = registerKey("end_xenoverd_ore_placed");
+
+    public static void bootstrap(BootstrapContext<PlacedFeature> context) {
+
+
+        var configuredFeatures = context.lookup(Registries.CONFIGURED_FEATURE);
+
+        register(context, XENOVERD_ORE_PLACED_KEY, configuredFeatures.getOrThrow(ModConfiguredFeatures.OVERWORLD_XENOVERD_ORE_KEY),
+                ModOrePlacement.commonOrePlacement(10, HeightRangePlacement.triangle(VerticalAnchor.absolute(-64), VerticalAnchor.absolute(60))));
+        register(context, END_XENOVERD_ORE_PLACED_KEY, configuredFeatures.getOrThrow(ModConfiguredFeatures.END_XENOVERD_ORE_KEY),
+                ModOrePlacement.commonOrePlacement(10, HeightRangePlacement.triangle(VerticalAnchor.absolute(-64), VerticalAnchor.absolute(60))));
+    }
+
+    private static ResourceKey<PlacedFeature> registerKey(String name) {
+        return ResourceKey.create(Registries.PLACED_FEATURE, ResourceLocation.fromNamespaceAndPath(Gregtech_pack_core.MODID, name));
+    }
+
+    private static void register(BootstrapContext<PlacedFeature> context, ResourceKey<PlacedFeature> key, Holder<ConfiguredFeature<?, ?>> configuration,
+                                 List<PlacementModifier> modifiers) {
+        context.register(key, new PlacedFeature(configuration, List.copyOf(modifiers)));
+    }
+}